Add protection against failed queries

+++ db_queries.py
@@ -53,15 +53,19 @@ def edit_restaurant(id, new_name):
     session = DBSession()
     # Get the indicated restaurant by ID, and update the name
     target = session.query(Restaurant).filter_by(id = id).one()
-    target.name = new_name
-    # Commit and close session
-    session.commit()
+    # Protect from an invalid query
+    if target != []:
+        target.name = new_name
+        # Remember to commit!
+        session.commit()
+    # Whether or not the query succeeded, close the session
     session.close()
 
 def delete_restaurant(id):
     session = DBSession()
     # Same as in edit_restaurant - get by id, perform function, commit, close
     target = session.query(Restaurant).filter_by(id = id).one()
-    session.delete(target)
-    session.commit()
+    if target != []:
+        session.delete(target)
+        session.commit()
     session.close()
